Changes in CSS and HTML Structure
Changed the structure of the HTML to re-add semantic tags and nest the
navigation inside the header, and all the content in a div in the body.

Removed duplicate ID usage

Added unique ID to links below main heading in order to prevent styling
from affecting those links and also the navigation links.

Added a couple hr lines to separate content visually.

Fixed the navigation menu from being hidden in the new html structure.
I’m sure this still needs work. It feels like I did too much for it,
and it could probably be simplified.