Start of Media Query and usage of Arrow
Put an Arrow at the bottom of header and extended the size of header to
cover entire screen on mobile devices.

Started the use of media queries on a desktop first approach. I know
it’s not the most optimal way to do it, but for now during the learning
process I feel like this way is easier for me to learn.