Content Update / More Responsiveness
Added Reflow of Projects and My Life using Media Queries to make site
more responsive.

Added my own content instead of place holders

Changed Header and Footer backgrounds.

Finally managed to fix the overlay on the Header to add a dark overlay
to make the text stand out. Did this using an alpha channel on a div,
and the z-index property.